Updated documentation referring to `BaseTopic`
In the previous commit, I renamed `Topic.DerivedTopic` to `Topic.BaseTopic` to better clarify the relationship (1006a2b). As part of that, I updated the documentation for the impacted members and any method parameters.

In addition, however, there are a number of cases where the documentation alludes to this relationship, even though the immediate member may not make a direct call to `Topic.BaseTopic`. These needed to be gone through carefully, as the word "derived" is often used for cases where it is still the most appropriate term (e.g., in terms of _derived classes_ or _derived types_ or _derived methods_, as used with regard to class inheritance and polymorphism).

This commit picks up those references.