Renamed `DerivedTopic` attribute descriptor to `BaseTopic`
During the database migration, rename the actual `TopicReferenceAttributeDescriptor` used for `DerivedTopic` to reflect the new name. In OnTopic 4.x, the name of the attribute was incidental to how it was stored, due to a quirk of the implementation, as well as historical conventions specific to `DerivedTopic`.

In OnTopic 5.x, that will be made consistent with other attributes, where the descriptor's key denotes the storage key (in this case, `ReferenceKey`).

As part of this, I updated the inline comments to better document what's going on, and to separate it from the topic references update that the original query was part of.