Introduced `TrackedItem<T>` as an abstraction for `AttributeValue`
The idea of the `AttributeValue` is to provide a key/value pair with metadata for tracking whether a record which is intended to be saved in the data store `IsDirty` and when it was `LastModified`.

This concept, however, isn't specific to attributes. It also pertains to topic references—or, at least, should. Given that, the new `TrackedItem<T>` provides an `abstract` class for handling both `AttributeValue` and a forthcoming `TopicReference` record.

Note: The names of derived records may change, but the concepts remain.