Updated `AttributeValue` to derive from `TrackedItem<T>`
The `TrackedItem<T>` class provides a base abstraction for centralizing common elements from `AttributeValue` so they can be shared with a forthcoming e.g. `TopicReference` record (4ed8ef1).

As the first step of implementing this, the `AttributeValue` record is being updated to derive from `TrackedItem<T>`—and, specifically, `TrackItem<string>`, since the value of an `AttributeValue` is a string.