Introduced empty constructor for use with generics
In a subsequent update, we'll be generalizing the format of `AttributeValueCollection`, which will necessitate creating `TrackedItem<T>`s without knowing anything about the _specific_ type. Unfortunately, a generic constraint can require an empty constructor, but not a constructor with a specific signature.

To support this, an empty constructor is introduced—and, by necessity, the otherwise required `Key` property is moved to support nulls, while being annotated with `[NotNull]`. That suggests that while the property _can_ be null after _construction_, it should not be null after _initialization_.