Derived `AttributeValueCollection` from the new `TrackedCollection`
The `TrackedCollection` offers a generic foundation for `AttributeValueCollection` (995b6fe). As it centralizes much of the functionality, that allows us to remove most of this functionality from the `AttributeValueCollection`.

The two pieces that remain needed are a) an overload to `SetValue()` that accepts `isExtendedAttribute`, and b) an overload of `IsDirty` that allows the exclusion of `LastModified` attributes (which are expected to change every update, and thus may not be an accurate reflection of actual changes to the data).

There are a couple of breaking changes due to how the overloads work. Most notably, the overload of `SetValue()` that allows disabling `enforceBusinessLogic` no longer supports the `isExtendedAttribute`. That's because the `enforceBusinessLogic` overload is only needed when making a call from a property setter (via `[AttributeSetter]`), and won't have any need or awareness of `isExtendedAttribute`; it should instead simply maintain whatever preexisting value exists. In practice, this should only affects unit tests and the `AttributeValueCollection` extension methods.

In addition, the internal `DeletedAttributes` property was renamed to `DeletedItems` to help generalize the functionality.

Finally, the default value for `GetValue()` was changed from empty (`""`) to `null`. This isn't a preference, but there's no easy way to guarantee not null with with the generic version, unfortunately, so it's better to be explicit by requiring the "" as a default value if that's important.