Updated calls to `GetValue()` to accept `null` default
Previously, `AttributeValueCollection.GetValue()` returned an empty string as the default. This is difficult to maintain with any guarantee of `[NotNull]` while generalizing the base `TrackedCollection`. As a result, if `[NotNull]` is needed, callers should now explicitly pass "" as the `defaultValue`, or otherwise handle potential `null` returns. This update corrects callers to make one of these two adjustments.