Removed explicit setting of `enforceBusinessLogic`
Previously, the `enforceBusinessLogic` parameter was explicitly set. The parameter order of the internal `SetValue()` overload has changed such that the order of the last two parameters would need to be flipped. But, there's no reason to call the overload here; the public overloads make more sense. In that case, simply removing the `enforceBusinessLogic` parameter ensures the internal overload is called, and that the `enforceBusinessLogic` continues to default to `true`.