Reverted the parameter order of `enforceBusinessLogic`
There are reasonable arguments for changing the parameter order of the internal `SetValue()` overload—as I did during the generalization of the `AttributeValueCollection` into `TrackedCollection` (995b6fe)—but, ultimately, the gains are marginal while the confusion is more significant, since a) the difference might not be registered by the compiler due to the two parameters both being Booleans (assuming the `isDirty` value isn't nullable), and b) this has long been the current order, and is thus familiar to internal developers.

Ultimately, it doesn't matter _too_ much, since it's an `internal` method, but to avoid potential confusion I'd rather keep it as is.