Derived `TopicReferenceCollection` from the new `TrackedCollection`
The `TrackedCollection` offers a generic foundation for `TopicReferenceCollection` (995b6fe). As it centralizes much of the functionality, that allows us to remove much of this functionality from the `TopicReferenceCollection`.

The only pieces that remain needed are a) the `IsFullyLoaded` property for tracking whether references were able to be merged into the local topic graph during load, and b) the protect overrides (i.e.., `InsertItem()`, `SetItem()`, `RemoveItem()`, and `ClearItems()`) so that we can handle the recipricol relationships (via `Topic.IncomingRelationships`) that's specific to the topic references, but not other tracked collections.

There are a couple of breaking changes due to generalizing the nomenclature. Most notably, `GetTopic()` and `SetTopic()` have been renamed to `GetValue()` and `SetValue()`. As this refers specifically to the `TrackedItem<T>.Value` field, that isn't unreasonable—though it does remove the hint that this is setting or getting a `Topic` specifically.

Finally, as this now derives from `KeyedCollection<>` instead of implementing `IDictionary<>`, a number of dictionary-specific methods have been removed, and the semantics are updated slightly. Most notably, the indexer no longer supports setting the value. As part of this, I also renamed it from `TopicReferenceDictionary` to `TopicReferenceCollection` to better articulate this change in semantics.

As the `TopicReferenceCollection` is new to OnTopic 5.x, these changes aren't expected to have any impact on downstream implementations—outside of OnTopic Editor, which is working off a preview release of OnTopic 5.x—and thus shouldn't have any impact on customers.

The next few commits will resolve the breaking changes within the OnTopic library itself.