Updated `Add()` to `SetValue()`
Since `TopicReferenceCollection` is no longer a dictionary (74d5907), the `Add()` overload doesn't make as much sense. This was required to fulfill the `IDictionary<>` contract, but is largely redundant with `SetValue()` with the `TrackedCollection`. We could easily just add this as a proxy function, but `SetValue()` is more consistent with the broader semantics of the library, and especially since we don't want callers to need to worry about whether or not a record already exists. As such, calls to `Add()` have been renamed to `SetValue()`.

Note: There is one exception to this. In the rare case that we want to add a new `TopicReference` class directly, we can still use the `Add()` method from `KeyedCollection`. This isn't normally preferred, as it's more cumbersome for callers to work with `TopicReference` records directly, instead of relying on `SetValue()` to handle them. But this is used as part of the `TopicReferenceCollection` unit tests for evaluating the business logic enforcement.