Updated XML Docs previously referring to `TopicReferenceDictionary`
Between `TopicReferenceDictionary` being renamed to `TopicReferenceCollection` and being updated to derive from `TrackedCollection<>`, all XML Documentation referring to the `TopicReferenceDictionary` is now broken. The documentation is now updated to either point to `TopicReferenceCollection` or, if pointing to a derived member, the underlying `TrackedCollection<>`. This required quite a bit of (re)wrapping due to the longer fully-qualified name of the generic base collection.