Replaced setter with new `TopicReference` instance
The new `TopicReferenceCollection`, now deriving from `TrackedCollection<>`, no longer supports setting a topic value via the indexer. We could replace this with `SetValue()`, as we did elsewhere. But, since this unit test is evaluating back doors, it makes more sense to inject a new `TopicReference` instance directly via `Add()`.

In production scenarios, this is almost never a preferred option, since working with the `TopicReference` record directly is a bit cumbersome; using `SetValue()` is much preferred. But since this is deliberately testing backdoors that might evade the logic of `SetValue()`, this is effective for the purpose of the unit test.