Updated expectations of unit test
Previously, setting a reference to `null` would remove it. That makes good sense. But now that the records represent `TopicReference` instances (with e.g. `IsDirty` and `Version`), it makes more sense to _update_ them to have a `null` `Value` instead of removing them, as that allows them to be tracked as `IsDirty`. Given that, I've updated the test of `RemoveItem()` to expect that the value still still be there—but then to verify that it's `null`.

(Strictly speaking, this isn't necessary in that they are _also_ tracked via `DeletedItems`, but it also doesn't hurt.)