Replace `TryGetValue()` with `GetValue()`
The `TopicReferenceDictionary` had a `TryGetValue()` method, since it implemented `IDictionary<>`. The `KeyedCollection<>` used by the new underlying `TrackedCollection` doesn't implement this. We could add one, but it doesn't add significantly to the capabilities of `GetValue()`, which wil check to see if a value exists, and if not return an optional default value. As such, the call to `TryGetValue()` is replaced with a call to `GetValue()`.

(I refactored this a bit since another condition relies on the variable, and thus it needs to be defined in the parent scope. This is done when placing `TryGetValue()` instead a condition's _expression_, but wouldn't be done if I declared the variable within the _body_ of the condition.)