Avoid infinite loop with `BaseTopic` getter
`BaseTopic` is a topic reference. Topic references can be inherited from the base topic. As such, calling `BaseTopic` calls `GetValue()` which in turn calls `BaseTopic`, causing an infinite loop. This isn't usually a challenge; it's a unique issue specific to the fact that `BaseTopic` is explicitly called by `GetValue()`. It can be easily fixed by manually retrieving the `BaseTopic` reference, if it exists, from the `TopicReferenceCollection` instead of relying on the (generally preferred) `GetValue()` method.