Updated `TopicPropertyDispatcher` to operate off of `TopicReference.Value`
The `TopicPropertyDispatcher` has custom handling for `AttributeValue` instances so that it can call the value from the `Value` object, while storing the `AttributeValue` object. Similar logic was needed for `TopicReference`, since `TopicReferenceCollection` now stores a wrapper object with metadata, instead of just a `Topic` reference.

The implementation is slightly different as `HasSettableProperty` needs a reference to the target type (i.e., `Topic`), whereas `AttributeValue` didn't since it was relying on the underlying `TypeMemberInfoCollection` to use its built-in conversion of basic scalar values to handle this.

This is a bit clumsy in that the `TopicPropertyDispatcher` only supports `AttributeValueCollection` and `TopicReferenceCollection`. As a result, having hard-coded exceptions for each of the primary customers isn't a great design. Ideally, this would be replaced with support for generic handling of e.g. `TrackedTopic`. As this is an internal library, that's not a priority, and the currenct approach is an easy way to maintain this functionality.