Generalized `TopicPropertyDispatcher<>` for `TrackedItem<>`
Since `TopicPropertyDispatcher<>` is exclusively used by `TrackedCollection<>` now, and `TrackedCollection<>` exclusively works with `TrackedItem<>` instances, we can update `TopicPropertyDispatcher` to be more intelligent by making its generics aware of `TrackedItem<>` and its `TValue` type. This allows us to get rid of a number of ugly conditions which checked to see if we were working with an `AttributeValue` or a `TopicReference` and respond accordingly. Now, instead, it just operates off of the underlying `TrackedItem<>`, thus providing a uniform treatment.

As this required adding a new generic type parameter, I used the opportunity to a) standardize the generic type argument order between `TopicPropertyDispatcher<>` and `TrackedCollection<>`, and b) rewrap the comments to account for the longer names, while also using a more compact format to help reduce unnecessary wrapping.