Renamed `IRelatedTopicBindingModel` to `IAssociatedTopicBindingModel`
The `RelatedTopicBindingModel` is intended to ensure that topic binding models can refer to related topics and have those relationships persisted. With the introduction of topic references, however, we've reevaluated the "relationship" nomenclature, since a related topic could be a relationship or a reference. The preferred terminology for referring to _both_ of these is "associations".

Given that, we've introduced a new `IAssociatedTopicBindingModel` interface and a new `AssociatedTopicBindingModel` implementation.