Bypass `ModelType.Reference` as part of `GetUnmatchedAttributes()`
Previously, `GetUnmatchedAttributes()` ignored `AttributeDescriptor`s on the source `ContentTypeDescriptor` with a `ModelType` of `Relationship` or `NestedTopic` since those aren't actually stored as attributes. This failed to include `Reference` types, however.

As a result, this meant that any topics with a topic reference attribute would always pass those topic references as null attributes to ensure they were deleted. Technically, this doesn't hurt anything—but it's also entirely unnecessary.

To mitigate this, the `ModelType.Reference` is included as part of this condition.

This breaks a unit test. This isn't because the underlying functionality stopped working, but because the content type being evaluated—`ContentTypeDescriptor`—only had one actual attribute—`Title`—and that was deliberately being set in the unit test. Previously, this failed to be recognized, however, because `ContentTypeDescriptor` had a `BaseTopic` reference. With the exclusion of `ModelType.Reference`, that was no longer being returned. This is easily fixed by updating the unit test to evaluate a `Page` content type instead, which has additional attributes defined.