Update `IsDirty` unit tests to operate off of "saved" topics
With the previous update, I prevented attributes, references, and relationships from being marked as clean if the topic was new, since a new topic is, by definition, dirty. (The one exception to this being the removal of an item.) This means that the unit tests for evaluating `IsDirty` needed to be updated to be "saved" entities—i.e., topics with an `Id`.