Applied `ITrackDirtyKeys` to `Topic`
To help enforce consistency between classes, the `ITrackDirtyKeys` interface has been applied to `Topic`, thus ensuring it can be interacted with in the same ways as other classes that provide tracking of dirty keys, such as `TopicReferenceCollection`, `AttributeValueCollection`, and `TopicRelationshipMultiMap`.

This required some juggling of the overloads to ensure that the contract was adhered to while still supporting extended scenarios not covered by the interface.