Added in `IsNew` handling for `Topic.IsDirty()`, `MarkClean()`
As with the collections (85c386f), a `Topic` that `IsNew` is, by definition, dirty—and, therefore, should not be permitted to be marked clean via e.g. `MarkClean()`.

Technically, there isn't any need to check `IsNew` on `IsDirty()`, since the topic cannot be `MarkClean()'ed if it `IsNew`. That said, this is a bit faster than doing a lookup against the `DirtyKeyCollection` if it's guaranteed to be dirty.

Note that this shortcut _only_ applies to `Topic`, and not its collections, since a new `Topic` _will_ have dirty attributes, but a `Topic`'s collections may be empty—and, thus, clean.