Mark `Parent` as dirty during update
Previously, there was no tracking of `IsDirty` for resetting the `Parent`. This was intentional, since changes to the parent are handled separated from changes to other properties—i.e., via `ITopicRepository.Move()`, not `ITopicRepository.Save()`—and, as such, we didn't want _moving_ a `Topic` to mark it as needing to be _saved_.

This prevented `Save()`, however, from detecting if a `Topic` needed to _also_ be `Move()`d as part of the operation, a feature we've long supported. This is, admittedly, a rare use case, but can come in handy when programmatically working with the topic graph—e.g., when making multiple changes, then doing a recursive `Save()`.

The actual implementation of this tracking will be implemented in a subsequent commit.