Implement `Move()` on `Save()` if `Parent` has changed
With the newly (re)implemented change tracking of `Parent` (ebd6d3c), the `TopicRepository` is able to fix the implementation of committing a `Move()` on `Save()` if the `Parent` has been modified.

This is a rare scenario, as typically a `Topic` will be moved via `ITopicRepository.Move()` and _then_ saved via `ITopicRepository.Save()`. In programmatic updates to the graph, however, it's conceivable that the parent will be modified _without_ calling `Move()`, in which case that update should be committed as part of `Save()`. This is a logical assumption from the implementations perspective, and thus helps ensure the integrity of the graph.