Introduce `IEnumerable<Topic>` extension methods
The `TopicCollectionExtension` methods provide some basic extension methods for evaluating all topics in an `IEnumerable<Topic>` collection. These are implemented as extensions so that they don't need to be separately defined on e.g. `TopicCollection`, `KeyedTopicCollection<T>`, &c. This allows these methods to be applied to all of those at once—assuming, of course, that `OnTopic.Querying` is in scope.

To start with, two extension methods are introduced. One will determine if any of the topics `IsDirty`. The other will determine if any of the topics `IsNew`.

These are pretty trivial shortcuts. We could implement this via a delegate as well—but, at that point, callers might as well simply call LINQ's `Any()` directly.