Prevent `MarkClean()` if any related items are `Topic.IsNew`
A relationship should not be permitted to be marked as clean if any of the related topics `IsNew`; that fact suggests that the relationship is not, in fact, clean (since not all items in it have been saved). This parallels the similar implementation on `TrackedCollection<>` (b0787c6), and helps make `MarkClean()` more intelligent.