Unit test: Prevent clean `TrackedItem<T>` being added to `Topic.IsNew`
If a `Topic` `IsNew` then any modifications to it should be marked as `IsDirty`, regardless of whether or not the caller requested that they be marked as such. This prevents scenarios where attributes are orphaned because they were marked as clean even though the topic hadn't yet been saved. This test evaluates the backdoor of inserting a new `AttributeValue` with `IsDirty` explicitly set to `false`.