Unit test: Prevent `!markDirty` being set on `Topic.IsNew`
If a `Topic` `IsNew` then any modifications to its relationships should be marked as `IsDirty`, regardless of whether or not the caller requested that they be marked as such. This prevents scenarios where relationships are orphaned because they were marked as clean even though the parent topic hadn't yet been saved.