Unit test: Prevent `!markDirty` being set with `Topic.IsNew` target
If a new `Topic` is being added to a `Topic.Relationships`, then it should not be allowed to be marked as `!IsDirty`, regardless of whether or not the caller requested that. This prevents scenarios where relationships are orphaned because they were marked as clean even though the target topic hadn't yet been saved.