Unit test: Prevent `!markDirty` being set with `Topic.IsNew` target
If a new `Topic` is being added to a `Topic.References`, then it should not be allowed to be marked as `!IsDirty`, regardless of whether or not the caller requested that. This prevents scenarios where references are orphaned because they were marked as clean even though the target topic hadn't yet been saved.

This is similar to a unit test on the `AttributeValueCollection` tests (dc08f2e), except that it evaluates whether the _target_ value `IsNew`, not the `AssociatedTopic`.