Implemented `TABLOCK` for the `CreateTopic` stored procedure
While we generally try to avoid aggressive locks such as `TABLOCK`, it becomes important when working with the nested set since minor operations—such as inserting a single record in this case—impact all records after the insert, due to their `RangeRight` and possibly `RangeLeft` being modified. Given that, we want to prevent any concurrent writes or even reads from occuring during this time, as otherwise they could end up with an inconsistent hierarchical state.

Fortunately, writes to the hierarchy are generally pretty uncommon—and OnTopic is broadly optimized for read operations, right down to the decision to use a nested set hierarchy in the first place.