Refactored topic collection list as graph
The purpose of the various topic collections is fairly self-explanatory based on their names, so having a description for each didn't add much value—while organizing them as a list made the various dimensions confusing. To mitigate that, I've refactored the topic collection list as a grid contrasting read-write to read-only against unkeyed, keyed, and generic collections. This is easier to read.