Removed references to legacy projects, added references to new projects
We're no longer actively maintaining the ASP.NET Web Forms or ASP.NET MVC Framework versions of OnTopic, and they will **not** work with OnTopic 5.x. As such, references to these projects are removed.

While I was at it, introduced references to new projects such as the new SQL unit tests and the Data Transfer library.