Reviewed, updated `README` for `OnTopic.Mapping.Hierarchical` namespace
Clarified language, expanded on concepts. As there are no major updates to the `HierarchicalTopicMappingService` in OnTopic 5.0.0, most of these changes are minor, and simply provide emphasis or more consistency in wording.

The biggest update was a bug fix in the sample code which excluded topics of the content type `PageGroup`, instead of topics whose _parent_ are of the content type `PageGroup`. This reflects, in turn, a bug fix which enforces this logic in OnTopic 5.0.0, whereas previously the validation function inadvertently ran on the parent topic, not each individual topic.