Reviewed, updated `README` for `OnTopic.Data.Caching` namespace
Clarified language, expanded on concepts. As there are no major updates to the `CachedTopicRepository` in OnTopic 5.0.0, most of these changes are minor, and simply provide emphasis or more consistency in wording. In addition to changing the version number for the package, I also emphasized that this decorator is recommended for web applications, in particular.