Reviewed, updated `README` for `OnTopic.AspNetCore.Mvc` namespace
Clarified language, expanded on concepts. Where there are no major updates to the `OnTopic.AspNetCore.Mvc` project in OnTopic 5.0.0, there were quite a few updates from OnTopic 4.x that weren't fully documented, and are now included.

This includes, for example, the `PageLevelNavigationViewComponent`, the `TopicRouteValueTransformer`, as well as the `IEndpointRouteBuilder` extension methods for handling areas, such as `MapTopicAreaRoute()` and `MapImplicitAreaControllerRoute()`.