Reviewed, updated `README` for `OnTopic.ViewModels` namespace
Clarified language, expanded on concepts. As there are no major updates to the `OnTopic.ViewModels` project in OnTopic 5.0.0, most of these changes are minor, and simply provide emphasis or more consistency in wording. In addition to changing the version number for the package, I also expanded on the behavior of the `DynamicTopicViewModelLookupService`, and preferred the more accurate term "parameterless constructor" to "default constructor".