Reviewed, updated `README` for `OnTopic.Data.Sql` namespace
Clarified language, expanded on concepts. As there are no major updates to the public interface of the `OnTopic.Data.Sql` project in OnTopic 5.0.0, most of these changes are minor, and simply provide emphasis or more consistency in wording. In addition to changing the version number for the package, I also clarified that the caching applies both to read and write operations.