Reviewed, updated `README` for `OnTopic.TestDoubles` namespace
There haven't been any real updates to the `TestDoubles` as part of OnTopic 5.0.0. That said, there were gaps in the documentation. Most notably, I documented the full structure of the `StubTopicRepository`, which will make it easier to work with. I also introduced a note that this package isn't intended for customer implementation—though, of course, nothing prevents customers from using it, if it's needed.