Rename `AttributeValue` to `AttributeRecord`
This aligns with the rename of `TrackedItem` to `TrackedRecord` (a6db252). This is a major breaking change, since `AttributeValue` is a long established class. That said, this avoids the awkward convention of e.g. `AttributeValue.Value`, and unifies the naming convention for combining metadata and values as a "Record".

(As it so happens, this also _happens_ to be a C# 9.0 record type—though that's not why we're using this naming convention here.)