Migrated `PropertyConfiguration` to `internal`
The `PropertyConfiguration` class has always been intended for `internal` use—as indicate by its placement in the `OnTopic.Mapping.Internal` namespace!—but it maintained a `public` access modifier because it was referenced as a parameter in `protected` methods off of the `TopicMappingService`.

Those methods have recently been moved to `private` (f74e2c9b) in order to acknowledge that they don't offer a fully thought out opportunity for extensibility. Given that, we can now migrate `PropertyConfiguration` to a _truly_ `internal` class.