Fixed references to `SetPropertyValue()`
In a previous commit, I merged two overloads of `SetPropertyValue()` into one which accepts a final parameter of `object?`, not `string`). To accommodate this, the XML Doc references to it needed to be updated. While I was at it, I rewrapped the XML Docs within the `TypeMemberInfoCollection` unit test class.