Marked all deprecations as obsolete
As this is a major version with a lot of breaking changes, we're going to tear off the bandaid and force adopters to utilize the new methods. That will also allow us to remove these deprecated methods in e.g. OnTopic 5.1.0, instead of carrying them around until e.g. OnTopic 6.0.0, which will likely be another year or two away.