Remove `virtual` from read-only properties
Instead of `virtual` properties, prefer `sealed` properties with a `protected init` accessor. This way, derived classes—and _only_ derived classes—can still override the _value_, but without needing to `override` the _property_.

In practice, this ends up not being that much less code, so the gain is small. Nevertheless, overriding properties introduces potential issues with polymorphism (e.g., if a value is set to a local backing field, not relayed through the `base` implementation). And since the only purpose for marking real-only fields as `virtual` is to modify the default value, there isn't a strong case for doing this now that we have `init` accessors.

(When these were introduced, we didn't have `init` accessors, though we could have used `protected set` instead to get a similar functionality.)