Removed unnecessary `[AttributeSetter]` attributes
The `[AttributeSetter]` attribute is used to ensure that business logic is enforced when setting attributes that have associated properties. This introduces overhead, and reflection is used to call the property.

While this is pretty well optimized for comparatively rare write scenarios, it nonetheless slows down the operation. That's fine if it's needed, but these properties are effectively just convenience pass-throughs that don't provide any data validation or state management, but simply provide a strongly typed property for accessing the underlying attribute.

Given that, removing the `[AttributeSetter]` doesn't risk a "backdoor" around enforcing the business logic, but speeds up the constructions of new `Topic`s.

Obviously, in the unlikely event that business logic is introduced into these properties, the `[AttributeSetter]` attribute will need to be reintroduced.