Ensured `isDirty` is optional on deprecated `SetTopic()` method
The `SetTopic()` method is deprecated, and should not be called. That said, the original version had an optional `isDirty` parameter, which many callers took advantage of. By reverting this parameter to be optional, that ensures that those callers received the `[Obsolete()]` message, whereas otherwise they get an unintuitive message that the overload doesn't exist. This is simply intended to aid in the migration process.