Implemented `DerivedTopic` as `[Obsolete()]` for warning messages
While `DerivedTopic` is obsolete, it had been removed entirely, thus preventing callers from receiving clear guidance from code analysis or the compiler as to how to resolve this issue. Temporarily reintroducing it as an `[Obsolete()]` member to provide direction. These obsolete members will likely be removed as part of OnTopic 5.1.0.