Corrected rename of `DerivedTopic`
While the C# property was named `DerivedTopic`, and the `AttributeKey` was named `TopicID`, the attribute name itself had a key of `InheritedTopic`. Confusing? That's part of why we've unified around `BaseTopic` as a consistent naming convention. Regardless, this upgrade script failed to work due to that.

In addition, the `ContentType` at this point is still `TopicReferenceAttribute`. The change to `TopicReferenceAttributeDescriptor` will occur in the subsequent script. This could be placed after that, but it doesn't really matter so long as it acknowledges the current state of the data.