Introduced dependency on `OnTopic.Data.Sql`
This is a more controversial decision as it assumes that implementors will be using a SQL Server backend. As that's the only production-ready `ITopicRepository` currently available, however, that is a safe assumption. Further, while there are non-production proof-of-concepts for e.g. MongoDB and Cosmos DB, they introduce significant limitations for e.g versioning, and aren't expected to be the primary client.