Introduced polyfill for `[MemberNotNull]` attribute
The `[MemberNotNull]` attribute is introduced in .NET 5, and not available in .NET Standard 2.1, which OnTopic targets. We could multi-target .NET 5 to gain access to it. But as it's one of the few .NET 5 capabilities we have an immediate need for, and it can easily be introduced via a polyfill, we're doing that. This allows us to annotate e.g. properties as a means of validating that local members—including private fields—will not be null after the annotated member is called.